feat variable names, divide utilities
