+++ routes/home.py
@@ -1,9 +1,7 @@
 from fastapi import APIRouter, HTTPException, File, UploadFile, Form
 from pydantic import BaseModel
-from utils import client, handle_exception
-from functools import wraps
-from bson import ObjectId
-from Models.home_model import BlockModel
+from utilities.utils import client, handle_exception
+from models.home_model import BlockModel
 
 app = APIRouter(tags=['Home'])
 mydb = client["Delit-test"]
@@ -43,20 +41,22 @@ async def get_all_blocks() -> list:
 @handle_exception
 async def get_block_data(name: str) -> dict:
     """
-    Retrieve data for a specific block by name.
-
-    This function takes the block name as input, converts it to lowercase, checks if it exists in the predefined blocks list,
-    and then retrieves the corresponding data from the database.
-
-    If the block is invalid or not found in the database, an appropriate HTTPException is raised.
+    Description:
+        
+        Retrieve data for a specific block by name.
+        This function takes the block name as input, converts it to lowercase, checks if it exists in the predefined blocks list,and then retrieves the corresponding data from the database.
+        If the block is invalid or not found in the database, an appropriate HTTPException is raised.
 
     Args:
+        
         name (str): The name of the block to retrieve.
 
     Returns:
+        
         dict: A dictionary containing the block's data (name, content, and image image_link).
 
     Raises:
+        
         HTTPException: If the block name is invalid or the data is not found in the database.
     """
     name = name.lower()
@@ -76,16 +76,19 @@ async def update_block(block_name: str, data: BlockModel):
     Update a specific block by its name.
 
     Args:
-    - block_name (str): The name of the block to update.
-    - data (BlockModel): The data to be updated.
+
+        block_name (str): The name of the block to update.
+        data (BlockModel): The data to be updated.
 
     Returns:
-    - result (dict): A dictionary containing the status of the update.
+
+        Result (dict): A dictionary containing the status of the update.
                     If the update is successful, it will contain {"success": "block updated successfully"}.
                     If the update is not successful, it will contain the error message.
 
     Raises:
-    - HTTPException: If the block name is invalid or the data is not found in the database.
+
+        HTTPException: If the block name is invalid or the data is not found in the database.
     """
     block_name = block_name.lower()
     result = await connection.find_one({"name": block_name})
@@ -104,15 +107,18 @@ async def delete_block(block_name: str):
     Delete a specific block by its name.
 
     Args:
-    - block_name (str): The name of the block to delete.
+
+        block_name (str): The name of the block to delete.
 
     Returns:
-    - result (dict): A dictionary containing the status of the deletion.
-                    If the deletion is successful, it will contain {"success": "block deleted successfully"}.
-                    If the deletion is not successful, it will contain the error message.
+
+        result (dict): A dictionary containing the status of the deletion.
+            If the deletion is successful, it will contain {"success": "block deleted successfully"}.
+            If the deletion is not successful, it will contain the error message.
 
     Raises:
-    - HTTPException: If the block name is invalid or the data is not found in the database.
+
+        HTTPException: If the block name is invalid or the data is not found in the database.
     """
     block_name = block_name.lower()
     result = await connection.find_one({"name": block_name})
