+++ utilities/gtihub_utilities.py
@@ -1,34 +1,9 @@
-from motor.motor_asyncio import AsyncIOMotorClient
-from pymongo.server_api import ServerApi
-from fastapi import HTTPException
-from functools import wraps
+from .utils import GITHUB_TOKEN, REPO_OWNER, REPO_NAME, FOLDER_PATH
+import httpx
 import datetime
 import base64
-import httpx
+from fastapi import HTTPException
 import re
-
-url = "mongodb+srv://phanivutla2004:phaniphani@cluster0.gddku.mongodb.net/myFirstDatabase?retryWrites=true&w=majority&ssl=true"
-client = AsyncIOMotorClient(url, server_api=ServerApi('1'),connectTimeoutMS = 50000)
-
-GITHUB_TOKEN = ""
-REPO_OWNER = "Harshad712"
-REPO_NAME = "RKV-SPORTS-API"
-FOLDER_PATH = "newfolder"
-
-
-def handle_exception(function):
-    @wraps(function)
-    async def wrapper(*arguments, **kwargs):
-        try:
-            return await function(*arguments, **kwargs)
-        except HTTPException as http_exce:
-            raise http_exce
-        except Exception as e:
-            raise HTTPException(
-                status_code=500, detail=f"An unknown error occurred.{str(e)}")
-
-    return wrapper
-
 
 async def upload_to_github(file_content, file_name):
     """ Uploading the actual image file into github repository
@@ -53,6 +28,7 @@ async def upload_to_github(file_content, file_name):
     }
     response = httpx.put(url, json=data, headers=headers)
     return response
+
 
 async def delete_file_from_github(link: str):
     """ Delete the file from github repository
